feat: panic
It's time to stop allowing gopkg.in/check code to land in 4.0. Instead,
we just need to put our hands up and call it a day on supporting that.

Instead, just panic - panic everywhere. This will force the merge
forwards to be fixed there and then, instead of me having to fix it
every time the go.mod file warns be about mistakes.